pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pdf2learn"
version = "0.1.0"
description = "Convert any PDF into a self-contained, learning-optimized HTML document."
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.10"
authors = [{ name = "3allafi" }]
keywords = ["pdf", "html", "learning", "marker-pdf", "converter"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Education",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Education",
    "Topic :: Text Processing :: Markup :: HTML",
]
dependencies = [
    "marker-pdf>=0.2",
    "jinja2>=3.1",
    "typer>=0.12",
    "rich>=13.0",
    "pyyaml>=6.0",
    "beautifulsoup4>=4.12",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.5",
    "reportlab>=4.0",
]

[project.scripts]
pdf2learn = "pdf2learn.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/pdf2learn"]

[tool.ruff]
line-length = 100
target-version = "py310"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "RUF"]
ignore = ["E501"]

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-ra --strict-markers"
